Add files via upload
Developed an expert system that combines artificial intelligence and cybersecurity to automate vulnerability detection, security analysis, and reporting. The system allows users to perform network and device security tests, analyze results using AI-driven techniques, and create detailed reports with specific recommendations. It includes a dynamic vulnerability database that updates automatically, a smart chatbot assistant that offers real-time cybersecurity support based on test results, and a role-based structure to manage administrative and developer permissions. By merging AI with penetration testing and threat analysis, the project improves the efficiency of cybersecurity professionals, lowers manual effort, and strengthens the overall security of targeted systems.

Many problems faced us when building this application, the first problem we encountered was testing and preparing an environment for vulnerability scanning and penetration testing (testing and deployment), the problems we faced is as listed:
1.	not enough knowledge about building networks in general and building vulnerable networks in specific.
2.	we tried for a real network but there was a lack of resources (“computers, routers, switches, cables, etc”). 
3.	after failing to build a real life network we opted for a virtual network
4.	most of the tools that build virtual network are for experts and was very hard to use and to navigate through 
5.	internet connection was always a problem so downloading vms and other machines was time consuming
6.	creating connections between machines and ensuring  this connections stays still was hard 
The next problem after building the network for testing and deployment was the automation of all of the process.

From the start, automating vulnerability scans and penetration testing when all tools for scanning and penetration testing require human intervention and interactions  that were hard to automate and required using low level languages to deal with it and be able to implement it with a high level language like python.
This raises another concern because dealing with low level languages and implementing them requires a lot of modifying and checking and dealing with low level errors that was very hard to fix and was very time consuming other than not having the enough knowledge and expertise to deal with such problems. 

implementing the full project with the mitre attack as a framework carried a lot of difficulties as follow:
1.	there are a lot of tools that u can use created by mitre themselves but have a very long learning curve and will not be easily implemented 
2.	the framework is very complex and needs a lot of time to understand how it works and functions
3.	if you want to make use of the database you have to download it or use a predefined lib that they created so that u can access its elements 
a.	if you decide to download it, this process will take a lot of time and disk space the size of the database is so big 
b.	If you decide to use the lib related to it, it's written in a special structure called (“stix”) and also needs a long learning curve to be able to understand how it works and make use of it.

The next and the most important problem was the mapping between CVE id’s and Mitre attack TTP.
the problems we encountered was as follows:
1.	there is no determined no doubt link between CVE ids and TTPs
2.	there is 3 ways to link CVE ids  to TTPs 
a.	the first one is mapping them one by one using logic and a human expertise 
i.	by the size of both NVD and Mitre attack database its impossible to link them by hand
b.	the second the using a prebuilt dataset that contains the mapping
i.	Very small size, so hard to find and the linking is not a 100 percent certainty that its correct
c.	The last way is to implement a pre-built machine learning or neural network that is trained on a small dataset learned the patterns and linked the rest or built to link by using nlp models and cosine similarity concept.

for our case we tried implementing several models the problem was some of them required older versions of libraries in python that when deployed created a lot of problems between the compatibilities of libraries with each other, and using NN models needed a lot of resources that we don’t have and required funding because parts of it used ChatGPT api keys so we couldn’t test it. 
for the first problem which is creating a virtual environment for the parts of vulnerability scanning and penetration testing when deploying the solution was to use VMs and creating a virtual network  it is build using 3 machines:
1.	Host “Windows 11 ”
2.	Attacker “Linux”
3.	Target (Vulnerable machine) “ Metasploitable 2”
4.	the host is linked to the attacker machine using a security ssh channel configured beforehand using public and private key exchange 
Creating this type of connection other than the high security for the connection between the host and the attacker it will help a lot during automation because either when we want to create an automated vulnerability scan or for the automation of the penetration testing all we want to send is the command we want to run on the attacker machine without the login process, without the password to identify authentication and authorization it will be automatically done using the shared keys between them.
To fix the problem of automation we built the application using modular structure, each module is created to achieve a goal and is implemented in the bigger module that needs it.