Create README.md
In recent days, with the rise of using artificial intelligence, increasing advancements in the field of software development and the constant demand for cloud computing services, the rate of using modular structure for building applications ranked up, although this new structure helped with terminating the single point of failure and increased usability for developers when constructing the application, it took a huge toll on the security part of the process because having everything built apart from each other  increased the need to protect each modular section of the application alone and even when implementing the required security on those sections the transactions of data between them on the network is not trusted to guarantee the safety of data from both passive and active attacks. 
Considering these new challenges, the tools for implementing security processes are very complicated and require a lot of human intervention, needing to have more professionals in the cybersecurity and network security field to ensure the safety of data either inside these modular structures or in the network in general, rapidly enlarged when a multitude of new challenges rose up, We don't have enough manpower of experts to be able to work in all the enterprises, from startups to cooperate and not enough liquidity, especially in the startup community, leaving them behind in the security aspect. 
To bridge the gap and give startups the chance to build their security without the worry of more expenses when hiring a cybersecurity analyst expert to do the job and boost the work efficiency and speed by helping experts in the big companies, our project Cybersecurity Analyst Expert System (“CyberXpert”) utilizes artificial intelligence strategies to automate the workflow of the Cybersecurity analyst from start to finish and implement a globally recognized, comprehensive knowledge base. Developed by the MITRE Corporation, MITRE ATT@CK Framework (“Adversarial Tactics, Techniques, and Common Knowledge”) as the backbone of the full project to build confidence in  it .